input: touchscreen: Add state_notifier driver
This module is used to notify touch driver about the current
screen status. It is useful for suspended processes in different drivers.
MDSS driver is a basis in this implementation.

Hammerhead uses LCD notifiers, so state_notifier is based on them.
It is easier to use this instead of LCD notifiers.

Signed-off-by: Alex Saiko <solcmdr@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>